@@ -1,39 +1,32 @@
 // Dashboard/Website/index.js
-import React, { Component } from 'react'
-import SideBar from '../Partials/SideBar'
-import S from 'shorti'
+import React from 'react'
 import config from '../../../../../config/public'
+import { connect } from 'react-redux'
 
 const API_URL = config.store.url
-
-export default class Website extends Component {
-  render() {
-    // Data
-    const data = this.props.data
-    const user = data.user
-    let brand = data.brand
-    let brand_id
-    if (!brand)
-      brand = user.brand
-    if (brand)
-      brand_id = brand.id
-    const main_style = S('absolute h-100p border-left-70-solid-fff w-100p')
-    return (
-      <div style={S('minw-1000')}>
-        <main>
-          <SideBar
-            data={data}
-          />
-          <div style={main_style}>
-            <iframe style={S('w-100p h-100p absolute')} src={`${API_URL}/store?access_token=${user.access_token}${brand_id ? `&brand=${brand_id}` : ''}`} frameBorder="0" />
-          </div>
-        </main>
-      </div>
-    )
-  }
+const Website = ({
+  user,
+  brand
+}) => {
+  let brand_id = brand || user.brand
+  return (
+    <main>
+      <iframe
+        title="Website builder"
+        style={{
+          position: 'absolute',
+          height: '100%',
+          width: 'calc(100% - 70px)'
+        }}
+        src={`${API_URL}/store?access_token=${user.access_token}${brand_id ? `&brand=${brand_id}` : ''}`}
+        frameBorder="0"
+      />
+    </main>
+  )
 }
 
-// PropTypes
-Website.propTypes = {
-  data: React.PropTypes.object
-}
+export default connect(({ data }) => ({
+  user: data.user,
+  brand: data.brand
+})
+)(Website)
