@@ -1,9 +1,8 @@
 import webpack from 'webpack'
-import path from 'path'
 import config from '../config/webpack'
 
 export default {
-  devtool: 'eval',
+  devtool: 'inline-source-map',
   entry: {},
   output: {
     path: config.compile.output,
