@@ -21,7 +21,7 @@ webpackConfig.entry = [
 webpackConfig.plugins.push(
   new webpack.HotModuleReplacementPlugin(),
   // new webpack.NoEmitOnErrorsPlugin(),
-  // new WebpackNotifierPlugin({ alwaysNotify: true })
+  new WebpackNotifierPlugin({ alwaysNotify: true })
 )
 
 webpackConfig.module.rules.push(
